Created specific buttons for answer choices and difficulty setting choices
