+++ src/components/StartPage/StartPage.js
@@ -2,24 +2,20 @@ import styles from './StartPage.module.css';
 import { useContext } from 'react';
 import { Context } from '../../Context';
 import { v4 as uuidv4 } from 'uuid';
-import ChoiceButton from '../ChoiceButton/ChoiceButton';
 import ActionButton from '../ActionButton/ActionButton';
+import DifficultyButton from '../SelectButtons/DifficultyButton';
 
 export default function StartPage() {
-  const { setDifficulty, handleStartGame } = useContext(Context);
+  const { handleStartGame } = useContext(Context);
   const difficultySettings = ['easy', 'medium', 'hard'];
 
-  const handleSetDifficulty = (e) => {
-    const choice = e.target.textContent;
-    setDifficulty(choice);
-  };
   return (
     <>
       <h1 className={styles.logo}>Trivia!</h1>
       <ul className={styles.choiceContainer}>
         {difficultySettings.map((setting) => (
           <li key={uuidv4()}>
-            <ChoiceButton label={setting} clickHandler={handleSetDifficulty} />
+            <DifficultyButton label={setting} />
           </li>
         ))}
       </ul>
